+++ src/biz/bokhorst/xprivacy/XWebView.java
@@ -82,7 +82,7 @@ public class XWebView extends XHook {
 				Class<?> clazz = param.getResult().getClass();
 				if (!mWebSettings.contains(clazz.getName())) {
 					mWebSettings.add(clazz.getName());
-					XPrivacy.hookAll(XWebSettings.getInstances(param.getResult()));
+					XPrivacy.hookAll(XWebSettings.getInstances(param.getResult()), getSecret());
 				}
 			}
 		} else