+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Test and beta releases will have experimental functions enabled by default.
 **Next release**
 
 * Added a switch to enable/disable all application restrictions
+* Prevent pollution of usage data ([issue](https://github.com/M66B/XPrivacy/issues/1140))
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)